updated telegram lib / int/int64 hell fixed / unified config retrieving

+++ util/array.go
@@ -1,7 +1,7 @@
 package util
 
 // IsIdInList Helper function to check if an ID is in a list of IDs
-func IsIdInList(id int, idList []int) bool {
+func IsIdInList(id int64, idList []int64) bool {
 	for _, listID := range idList {
 		if id == listID {
 			return true
